FixedAll

+++ BlogPostify.Domain/Entities/Like.cs
@@ -1,10 +1,11 @@
 ﻿using BlogPostify.Domain.Commons;
+using BlogPostify.Domain.Entities.Users;
 
 namespace BlogPostify.Domain.Entities;
 
 public class Like : Auditable<long>
 {
-    public long PostId { get; set; }
+    public int PostId { get; set; }
     public Post Post { get; set; }
 
     public int UserId { get; set; }
