+++ BlogPostify.Service/Interfaces/Posts/IPostService.cs
@@ -8,6 +8,6 @@ public interface IPostService
     Task<bool> RemoveAsync(int id);
     Task<PostForResultDto> RetrieveIdAsync(int id);
     Task<PostForResultDto> AddAsync(PostForCreationDto dto);
-    Task<PostForResultDto> ModifyAsync(long id,PostForUpdateDto dto);
+    Task<PostForResultDto> ModifyAsync(int id,PostForUpdateDto dto);
     Task<IEnumerable<PostForResultDto>> RetrieveAllAsync(PaginationParams @params);
 }
